Revert "TW-2182 Integrate react-native-sapling with downgraded airgap-sapling version"
This reverts commit 3f5c48d6d2f8667a47e223bc7cc4142a286f7cea.
