+++ src/utils/sapling/sapling-keys/in-memory-spending-key.ts
@@ -4,11 +4,11 @@ import {
   getExtendedSpendingKey,
   getExtendedFullViewingKeyFromSpendingKey,
   prepareSpendDescriptionWithSpendingKey,
-  signSpendDescription as nativeSignSpendDescription
+  signSpendDescription as nativeSignSpendDescription,
+  getProofAuthorizingKey
 } from 'react-native-sapling';
 
 import { bufToUint8Array, toBase64 } from '../helpers';
-import { getProofAuthorizingKey } from '../sapling-functions-supplement';
 import { ParametersSpendProof, ParametersSpendSig, SaplingSpendDescription, SaplingTransactionInput } from '../types';
 
 import { decryptKey } from './helpers';
